[omf_trace] Convert header fields to host byte order.
Header fields longer than 1 byte in IP, TCP and UDP have to be converted
to host byte order, otherwise, values are meaningless. Also, the IP
fragment offset header is only 13 bits, with the DF and MF flags in some
of the MSB. They are zeroed out (and ignored in the OML report) before
byte-order conversion.

Fixes #531

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>