[wattsup] Fix multiplier issues, and use format strings relevant to precision
Some of the metrics reported by the Watts Up? .net (e.g., power factor
or, more importantly, current) are not in tenth of the base unit.
However, the previous code had this multiplier hard-coded. This commit
introduces per-field multipliers, as well as a format string relevant to
the precision to induced by this conversion.

Note that this might highlight some errors in the  Communication
protocol documentation [0], in section IV “UNITS, FORMAT, AND RANGE
FOR THE VALUES SENT FROM THE DATA LOG”.
* “<A>” is listed as “<A> Integer, amps * 10. (Thousandths of amps)”,
  which is incoherent (should be * 1000 or Tenths); comparing output from
  the reader's LCD and the serial message, the correct description should
  be “<A> Integer, amps * 1000.  (Thousandths of amps)”.
* “<Cost>” is assumed to be using cents as the base unit in the code,
  which is coherent with [0], but not quite natural. The reader's LCD
  reports $.
* The relation between “<PF>” and “<VA>” could be clarified:
    The apparent power <VA> is voltage times current. With phase
    differences, only a limited amount of power <W> is actually useful.
    The relation between these two numbers is the power factor <PF> such
    that
      W = VA * PF
    with VA > W and 0 < pr < 1.

[0] https://www.wattsupmeters.com/secure/downloads/CommunicationsProtocol090824.pdf

refs: #1141

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>