[wattsup] Fix multiplier issues, and use format strings relevant to precision
Some of the metrics reported by the Watts Up? .net (e.g., apparent
power or, more importantly, current) are not in tenth of the base unit.
However, the previous code had this multiplier hard-coded. This commit
introduces per-field multipliers, as well as a format string relevant
to the precision to induced by this conversion.

Note that this might highlight some errors in the  Communication
protocol documentation [0], in section IV “UNITS, FORMAT, AND RANGE
FOR THE VALUES SENT FROM THE DATA LOG”.
* “<A>” is listed as “<A> Integer, amps * 10. (Thousandths of amps)”,
  which is incoherent (should be * 1000 or Tenths); comparing output from
  the reader's LCD and the serial message, the correct description should
  be “<A> Integer, amps * 1000.  (Thousandths of amps)”.
* “<Cost>” is assumed to be using cents as the base unit in the code,
  which is coherent with [0], but not quite natural. The reader's LCD
  reports $.
* “<PF>” is described as “percent ratio of Watts versus Volt Amp”,
  which is incorrect: it is just the ratio, reported in tenth of units;
  the description should read “<PF> Power factor, PF * 10. (Tenths of
  ratio), ratio of Watts versus Volt Amps”.
* “<VA>” is described as “VA * 10 (tenths of volt-amps)”, while
  observation and comparison with <W>/<PF> indicates that it is probably
  “Volt-Amps, VA * 100. (tenths of volt-amps)”.

[0] https://www.wattsupmeters.com/secure/downloads/CommunicationsProtocol090824.pdf

refs: #1141

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>